Refactor: python, cts: Move SpecialTest1 into the _cts module.
And rename it to ResyncCIB.  No other code changes.
