+++ cts/lab/CTStests.py
@@ -43,52 +43,7 @@ AllTestClasses.append(ResourceRecover)
 AllTestClasses.append(ComponentFail)
 AllTestClasses.append(SplitBrainTest)
 AllTestClasses.append(Reattach)
-
-
-class SpecialTest1(CTSTest):
-    '''Set up a custom test to cause quorum failure issues for Andrew'''
-    def __init__(self, cm):
-        CTSTest.__init__(self,cm)
-        self.name = "SpecialTest1"
-        self._startall = SimulStartLite(cm)
-        self.restart1 = RestartTest(cm)
-        self.stopall = SimulStopLite(cm)
-
-    def __call__(self, node):
-        '''Perform the 'SpecialTest1' test for Andrew. '''
-        self.incr("calls")
-
-        #        Shut down all the nodes...
-        ret = self.stopall(None)
-        if not ret:
-            return self.failure("Could not stop all nodes")
-
-        # Test config recovery when the other nodes come up
-        self._rsh(node, "rm -f " + BuildOptions.CIB_DIR + "/cib*")
-
-        #        Start the selected node
-        ret = self.restart1(node)
-        if not ret:
-            return self.failure("Could not start "+node)
-
-        #        Start all remaining nodes
-        ret = self._startall(None)
-        if not ret:
-            return self.failure("Could not start the remaining nodes")
-
-        return self.success()
-
-    @property
-    def errors_to_ignore(self):
-        """ Return list of errors which should be ignored """
-
-        # Errors that occur as a result of the CIB being wiped
-        return [ r"error.*: v1 patchset error, patch failed to apply: Application of an update diff failed",
-                 r"error.*: Resource start-up disabled since no STONITH resources have been defined",
-                 r"error.*: Either configure some or disable STONITH with the stonith-enabled option",
-                 r"error.*: NOTE: Clusters with shared data need STONITH to ensure data integrity" ]
-
-AllTestClasses.append(SpecialTest1)
+AllTestClasses.append(ResyncCIB)
 
 
 class NearQuorumPointTest(CTSTest):
